Update how-to-insert-a-picture-into-a-word-processing-document.md
The current code contains a bug that makes the document not open in MS Word. This single error (`https` instead of `http` in URI) causes the document to be unopenable in any tested Word product (Word for Office 365 Win/macOS/iOS, MS Word Professional 2016).

Interestingly it does open fine in LibreOffice and others.