Add samples projects to a solution
This starts the process of refactoring the samples to not be inline.
Instead, this change creates a solution that the samples are added to in
order to validate they are correct (and enable refactoring much easier
in the future). This adds a script in the samples directory that helps
scaffold out a new sample project.