Add docs output directory to .gitignore
docs/open-xml-docs is where the docfx build ends up - it appears to have out of date assets that are overwritten with the build. This change removes the directory and adds it to .gitignore

As part of this change, a default template is added to docfx.json so that local builds can work.